requirements.txt
streamlit
pandas
seaborn
matplotlib
xgboost
matplotlib
sklearn.ensemble
xgboost
shap
numpy
sklearn.preprocessing
sklearn.model_selection
plotly.express
sklearn.cluster
sklearn.metrics
sklearn.linear_model
sklearn.pipeline
scipy.interpolate